usb: move USB transfer functions to appropriate file
Move USB transfer functions to appropriate file.
